refactor: consolidate managed build lifecycle
