+++ lambda/environments/tools-status.js
@@ -6,58 +6,19 @@ import {
   ECRClient,
 } from '@aws-sdk/client-ecr';
 import { GetObjectCommand, S3Client } from '@aws-sdk/client-s3';
+import {
+  RETRYABLE_ECR_ERRORS,
+  createBuildLifecycleHandler,
+  isUnsupportedScan,
+  streamToString,
+  summarizeScanFindings,
+} from './build-lifecycle.js';
 import { createToolStore } from './tool-store.js';
 
 const ddb = DynamoDBDocumentClient.from(new DynamoDBClient({}));
 const ecr = new ECRClient({});
 const s3 = new S3Client({});
 const defaultStore = createToolStore({ ddb });
-
-const RETRYABLE_ECR_ERRORS = new Set([
-  'ImageNotFoundException',
-  'ScanNotFoundException',
-  'ServerException',
-  'ThrottlingException',
-  'TooManyRequestsException',
-]);
-
-const streamToString = async (body) => {
-  if (!body) return '';
-  if (typeof body.transformToString === 'function') return body.transformToString();
-  const chunks = [];
-  for await (const chunk of body) chunks.push(chunk);
-  return Buffer.concat(chunks).toString('utf8');
-};
-
-const findingAttribute = (finding, key) =>
-  finding.attributes?.find((attribute) => attribute.key === key)?.value ?? null;
-
-const summarizeFindings = (scan) => {
-  const order = new Map([
-    ['CRITICAL', 0],
-    ['HIGH', 1],
-    ['MEDIUM', 2],
-    ['LOW', 3],
-    ['INFORMATIONAL', 4],
-    ['UNDEFINED', 5],
-  ]);
-  return (scan.imageScanFindings?.findings ?? [])
-    .map((finding) => ({
-      id: finding.name ?? 'Unknown finding',
-      severity: finding.severity ?? 'UNDEFINED',
-      packageName: findingAttribute(finding, 'package_name'),
-      packageVersion: findingAttribute(finding, 'package_version'),
-      uri: finding.uri ?? null,
-    }))
-    .toSorted(
-      (left, right) =>
-        (order.get(left.severity) ?? 99) - (order.get(right.severity) ?? 99) ||
-        left.id.localeCompare(right.id),
-    );
-};
-
-const isUnsupportedScan = (status, description) =>
-  status === 'UNSUPPORTED_IMAGE' || /UnsupportedImageError/i.test(description ?? '');
 
 const failVersion = async (store, version, reason, detail = null, patch = {}) =>
   store.updateVersion(
@@ -198,7 +159,7 @@ const inspectToolImage = async ({ store, version, ecrClient = ecr, s3Client = s3
     const scanFindings = {
       status: scanStatus,
       severityCounts,
-      findings: summarizeFindings(scan),
+      findings: summarizeScanFindings(scan),
       findingsTruncated: Boolean(scan.nextToken),
       evaluatedAt: new Date().toISOString(),
       imageDigest: image.imageDigest,
@@ -294,22 +255,16 @@ const poll = async ({ store, ecrClient, s3Client }) => {
   return { results };
 };
 
-export const createToolsStatusHandler =
-  ({ store = defaultStore, ecrClient = ecr, s3Client = s3 } = {}) =>
-  async (event) => {
-    try {
-      if (event?.action === 'poll') return poll({ store, ecrClient, s3Client });
-      if (event?.source === 'aws.codebuild') {
-        return handleBuildEvent({ store, event, ecrClient, s3Client });
-      }
-      if (event?.source === 'aws.ecr') {
-        return handleScanEvent({ store, event, ecrClient, s3Client });
-      }
-      return { ignored: true };
-    } catch (error) {
-      console.error('Managed tool status handling failed:', error);
-      throw error;
-    }
-  };
+export const createToolsStatusHandler = ({
+  store = defaultStore,
+  ecrClient = ecr,
+  s3Client = s3,
+} = {}) =>
+  createBuildLifecycleHandler({
+    label: 'Managed tool',
+    poll: () => poll({ store, ecrClient, s3Client }),
+    handleBuildEvent: (event) => handleBuildEvent({ store, event, ecrClient, s3Client }),
+    handleScanEvent: (event) => handleScanEvent({ store, event, ecrClient, s3Client }),
+  });
 
 export const handler = createToolsStatusHandler();
